fix: rename environment folder and update imports
